Rework CanoPy to be contained within Canopy class, follow standard Python module file structure, and generate the required config itself without manual copying.
If the input config is not detected when the class is initialized then a .cfg file will be generated from the template specified in './canopy/templates/config_template.py'.
