+++ canopy/__init__.py
@@ -1,0 +1,1 @@
+from .canopy import Canopy
