+++ canopy/templates/__init__.py
@@ -1,0 +1,1 @@
+from .config_template import config_template
